When the user is viewing messages in a room which the administrator removes then he gets notified and the send button and the input area are being disabled. For this purpose, added the roomRemovedNotification function
Added a check to see if the input area and the send button is enabled when the user selects a room. If they are disabled they are being enabled.

Also fixed userInteractionEnabled (set to true) for the input area when having a connection after it has been lost so the user can insert text to send.