Added UIRefreshControl -
the spinning indicator when you drag the top of the tableview further down. When this happens a spinning indicator appears. Added an action to stop the spinning indicator when you release the tableview. The loading of the previous messages doesn't happen here cause we want to get past messages before the user gets to the top of the tableview.

Added function pullToRefresh so we can load messages and append them above the cell where this function was called.
Added function scrollViewDidEndDragging so we can start loading past messages before the user gets at the top. When the user pass the 10th row the we call pullToRefresh to append the messages.

Added action loadMore. This is connected to the button in the header of the tableview. When you tap on it, it loads 10 messages and appends them above the existing messages of the tableview.

Added some parameters to the loadHistory function so we can use it accordingly from where we are calling it. For example when we first load messages in an room we just want to load the latest messages but if we scroll at the top we want to append messages on top of the existing messages. More specific added end and ls parameters which are optionals and represent the beginning and the end of the range of the messages we want to retrieve. Parameter append is a boolean to check if the retrieved messages are going to be appended above existing messages and the appendAboveCell parameter is to get the row of the cell which we will append above the retrieved messages.
So the difference with the previous loadHistory is when we want to append the messages. In that case we get the messages in a temp array reverse it and then the tmp array plus the array with the existing messages are being merged to the chatMessageData dictionary. Lastly the tableview scrolls to the row of the appendAboveCell parameter so the user won't loose his position on the tableview. If there are no more messages then we just stop the indicator from spinning.

Of course now that loadHistory's parameter's are more we add the in each call of the loadHistory function