update so that users can change to other servers
right now, if a user logs into the mobile console, then logs out on a desktop (log out of all other devices), then there is no way for the user to make it back to the server in the mobile iOS instance.  enabling the button at all times gives people the option of being able to connect to a different server whenever they wish.